initial version of generate stage, IIL;  also added -stage
The new generate stage converts from the AST to a new Internal
Instruction Language.  It is running now but its output is currently
ignored.  It only produces grammar elements, not actions yet.

Added new -stage command line flag to limit how far we get for each file
in terms of stages.  Attempts to produce stage-relevant output for each
file; for example, for the transform stage it outputs the before and
after versions of the statements.